Nghi

+++ BLL/DTOs/EquipmentStatsDTO.cs
@@ -14,11 +14,11 @@ namespace BLL.DTOs
         public int BonusHp { get; set; }
         public int BonusAtk { get; set; }
         public int BonusDef { get; set; }
-        public int BonusMoveSpeed { get; set; }
-        public int BonusAttackSpeed { get; set; }
-        public int BonusCritRate { get; set; }
-        public int BonusCritDamage { get; set; }
-        public int BonusDamageBonus { get; set; }
+        public float BonusMoveSpeed { get; set; }
+        public float BonusAttackSpeed { get; set; }
+        public float BonusCritRate { get; set; }
+        public float BonusCritDamage { get; set; }
+        public float BonusDamageBonus { get; set; }
 
         public int TotalHp => BaseHp + BonusHp;
         public int TotalAtk => BaseAtk + BonusAtk;
